Keep EcoreTypeInfo configuration for `@JsonType` annotations
When a `@JsonType` annotation is used to configure a different
`property` to extract the type info, the `ValueReader` and `ValueWriter`
functions to (de)serialize the property's content can only be configured
using `use = "CLASS"` or `use = "NAME"`.

However, when configuring the `EMFModule` using a custom `EcoreTypeInfo`
object, one can specify arbitrary functions to (de)serialize the type.

This change will copy the default `EcoreTypeInfo` configuration of the
`EMFModule` to any `EcoreTypeInfo` instances that have been configured
using annotations in the model. This will not change anything for places
where a `use = ...` annotation detail is in place.

Closes https://github.com/eclipse-emfcloud/emfjson-jackson/issues/43